pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "catalog-intelligence-pipeline"
version = "0.1.0"
description = "Catalog Intelligence Pipeline (Vision + GenAI + GCP-ready stubs)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Catalog Intelligence Team" }]
dependencies = [
    "fastapi>=0.111.0",
    "typer>=0.12.5",
    "pydantic>=2.7.0",
    "uvicorn>=0.30.0",
    "requests>=2.32.0",
    "pillow>=10.3.0",
    "duckdb>=1.1.2",
    "jsonschema>=4.22.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.5.0", "mypy>=1.10.0", "pytest>=8.2.0"]

[project.scripts]
catalog-pipeline = "catalog_intelligence_pipeline.cli:app"

[tool.ruff]
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "PL"]
ignore = ["PLR2004", "B008", "PLR0913"]

[tool.mypy]
python_version = "3.11"
strict = false
check_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
